Update khanacademy_revealer.user.js
I was confused as to why this code wouldn't inject into chrome, and I worked on some things before realizing I needed to turn on Developer Mode. Changes:
- Improved error handling
- Cleaned up log logic
- Easier debugging
This code got me through highschool, and I'll be going to OSU for Aerospace Engineering in fall of 2025. Thank you, Alex Dubov.